Development do premaster

See merge request parp/aktywny_dyrektor!727

+++ src/ParpV1/MainBundle/Controller/ZastepstwoController.php
@@ -3,6 +3,7 @@
 namespace ParpV1\MainBundle\Controller;
 
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
@@ -13,6 +14,7 @@ use APY\DataGridBundle\Grid\Source\Entity;
 use APY\DataGridBundle\Grid\Column\ActionsColumn;
 use APY\DataGridBundle\Grid\Action\RowAction;
 use APY\DataGridBundle\Grid\Export\ExcelExport;
+use DateTime;
 use ParpV1\MainBundle\Entity\Zastepstwo;
 use ParpV1\MainBundle\Form\ZastepstwoType;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
@@ -28,12 +30,41 @@ class ZastepstwoController extends Controller
     /**
      * Lists all Zastepstwo entities.
      *
-     * @Route("/index", name="zastepstwo")
-     * @Template()
+     * @Route("/index/{aktywne}", name="zastepstwo", defaults={"aktywne" : true})
+     *
+     * @param bool $aktywne
+     *
+     * @return Response
      */
-    public function indexAction()
+    public function indexAction(bool $aktywne = true): Response
     {
         $source = new Entity(Zastepstwo::class);
+        $now = new DateTime();
+
+        $tableAlias = $source->getTableAlias();
+        $source->manipulateQuery(
+            function ($query) use ($tableAlias, $now, $aktywne) {
+                if ($aktywne) {
+                    $query->andWhere($tableAlias . '.dataDo >= :now')
+                        ->setParameter(':now', $now)
+                        ->addOrderBy($tableAlias . '.dataDo', 'ASC');
+                } else {
+                    $query->andWhere($tableAlias . '.dataDo < :now')
+                        ->setParameter(':now', $now)
+                        ->addOrderBy($tableAlias . '.dataDo', 'DESC');
+                }
+            }
+        );
+        $source->manipulateRow(
+            function ($row) use ($now) {
+                if ($row->getField('dataOd') > $now) {
+                    $row->setColor('#d6e8f2');
+                    $row->setField('opis', '<i class="fad fa-hourglass-half"></i> ' . $row->getField('opis'));
+                }
+
+                return $row;
+            }
+        );
 
         $grid = $this->get('grid');
         $grid->setSource($source);
@@ -47,25 +78,51 @@ class ZastepstwoController extends Controller
                 ->setFilterable(false)
                 ->setSafe(true);
 
+        // Podgląd zastępstwa
+        $rowAction1 = new RowAction('<i class="far fa-search"></i> Podgląd', 'zastepstwo_show');
+        $rowAction1->setColumn('akcje');
+        $rowAction1->addAttribute('class', 'btn btn-info btn-xs');
+
         // Edycja zastępstwa
         $rowAction2 = new RowAction('<i class="fas fa-pencil"></i> Edycja', 'zastepstwo_edit');
         $rowAction2->setColumn('akcje');
         $rowAction2->addAttribute('class', 'btn btn-success btn-xs');
+        $rowAction2->addManipulateRender(
+            function ($action, $row) use ($now) {
+                if ($row->getField('dataDo') > $now) {
+                    return $action;
+                } else {
+                    return null;
+                }
+            }
+        );
 
         // Usunięcie zastępstwa
         $rowAction3 = new RowAction('<i class="far fa-trash-alt"></i> Skasuj', 'zastepstwo_delete');
         $rowAction3->setColumn('akcje');
         $rowAction3->addAttribute('class', 'btn btn-danger btn-xs');
+        $rowAction3->addManipulateRender(
+            function ($action, $row) use ($now) {
+                if ($row->getField('dataOd') > $now) {
+                    return $action;
+                } else {
+                    return null;
+                }
+            }
+        );
 
+        $grid->addRowAction($rowAction1);
         $grid->addRowAction($rowAction2);
         $grid->addRowAction($rowAction3);
 
         $grid->addExport(new ExcelExport('Eksport do pliku', 'Plik'));
-
         $grid->isReadyForRedirect();
 
-        return $grid->getGridResponse();
+        return $grid->getGridResponse("ParpMainBundle:Zastepstwo:index.html.twig", [
+                'aktywne' => $aktywne
+        ]);
     }
+
     /**
      * Creates a new Zastepstwo entity.
      *
@@ -80,10 +137,11 @@ class ZastepstwoController extends Controller
 
         if ($form->isValid()) {
             $entityManager = $this->getDoctrine()->getManager();
+            $entity->setLastModifiedBy($this->getUser()->getUsername());
             $entityManager->persist($entity);
             $entityManager->flush();
 
-            $this->addFlash('warning', 'Zastepstwo zostało utworzone.');
+            $this->addFlash('warning', 'Zastępstwo zostało utworzone.');
 
             return $this->redirect($this->generateUrl('zastepstwo'));
         }
@@ -110,7 +168,7 @@ class ZastepstwoController extends Controller
             'method' => 'POST',
         ));
 
-        $form->add('submit', SubmitType::class, array('label' => 'Utwórz Zastepstwo', 'attr' => array('class' => 'btn btn-success' )));
+        $form->add('submit', SubmitType::class, array('label' => 'Dodaj zastępstwo', 'attr' => array('class' => 'btn btn-success' )));
 
         return $form;
     }
@@ -147,7 +205,7 @@ class ZastepstwoController extends Controller
         $entity = $em->getRepository(Zastepstwo::class)->find($id);
 
         if (!$entity) {
-            throw $this->createNotFoundException('Unable to find Zastepstwo entity.');
+            throw $this->createNotFoundException('Podane zastępstwo nie zostało odnalezione.');
         }
 
         $deleteForm = $this->createDeleteForm($id);
@@ -171,10 +229,13 @@ class ZastepstwoController extends Controller
         $entity = $entityManager->getRepository(Zastepstwo::class)->find($id);
 
         if (!$entity) {
-            throw $this->createNotFoundException('Unable to find Zastepstwo entity.');
+            throw $this->createNotFoundException('Podane zastępstwo nie zostało odnalezione.');
         }
 
-        if (!in_array("PARP_ADMIN", $this->getUser()->getRoles()) && !in_array("PARP_ADMIN_ZASTEPSTW", $this->getUser()->getRoles()) && $entity->getKogoZastepuje() != $this->getUser()->getUsername()) {
+        if (!in_array("PARP_ADMIN", $this->getUser()->getRoles())
+            && !in_array("PARP_ADMIN_ZASTEPSTW", $this->getUser()->getRoles())
+            && $entity->getKogoZastepuje() != $this->getUser()->getUsername()
+            ) {
             $this->addFlash('warning', 'Nie masz uprawnień do edycji nie swoich zastępstw.');
             return $this->redirect($this->generateUrl('zastepstwo'));
         }
@@ -198,6 +259,12 @@ class ZastepstwoController extends Controller
     */
     private function createEditForm(Zastepstwo $entity)
     {
+        $now = new DateTime();
+
+        if ($entity->getDataDo() < $now) {
+            $this->addFlash('danger', 'Nie można edytować zakończonych zastępstw. Wprowadzone zmiany nie zostaną zapisane.');
+        }
+
         $form = $this->createForm(ZastepstwoType::class, $entity, array(
             'current_user' => $this->getUser(),
             'ad_users' => $this->getUsersFromAD(),
@@ -218,12 +285,18 @@ class ZastepstwoController extends Controller
      */
     public function updateAction(Request $request, $id)
     {
-        $em = $this->getDoctrine()->getManager();
+        $entityManager = $this->getDoctrine()->getManager();
+        $now = new DateTime();
 
-        $entity = $em->getRepository(Zastepstwo::class)->find($id);
+        $entity = $entityManager->getRepository(Zastepstwo::class)->find($id);
 
         if (!$entity) {
-            throw $this->createNotFoundException('Unable to find Zastepstwo entity.');
+            throw $this->createNotFoundException('Podane zastępstwo nie zostało odnalezione.');
+        }
+
+        if ($entity->getDataDo() < $now) {
+            $this->addFlash('danger', 'Nie można edytować zakończonych zastępstw.');
+            return $this->redirectToRoute('zastepstwo');
         }
 
         $deleteForm = $this->createDeleteForm($id);
@@ -231,7 +304,8 @@ class ZastepstwoController extends Controller
         $editForm->handleRequest($request);
 
         if ($editForm->isValid()) {
-            $em->flush();
+            $entity->setLastModifiedBy($this->getUser()->getUsername());
+            $entityManager->flush();
             $this->addFlash('warning', 'Zmiany zostały zapisane');
             return $this->redirect($this->generateUrl('zastepstwo_edit', array('id' => $id)));
         }
@@ -252,16 +326,25 @@ class ZastepstwoController extends Controller
     {
         $form = $this->createDeleteForm($id);
         $form->handleRequest($request);
+        $now = new DateTime();
 
         if ($form->isValid()) {
             $entityManager = $this->getDoctrine()->getManager();
             $entity = $entityManager->getRepository(Zastepstwo::class)->find($id);
 
             if (!$entity) {
-                throw $this->createNotFoundException('Unable to find Zastepstwo entity.');
+                throw $this->createNotFoundException('Podane zastępstwo nie zostało odnalezione.');
             }
 
-            if (!in_array("PARP_ADMIN", $this->getUser()->getRoles()) && !in_array("PARP_ADMIN_ZASTEPSTW", $this->getUser()->getRoles()) && $entity->getKogoZastepuje() != $this->getUser()->getUsername()) {
+            if ($entity->getDataOd() < $now) {
+                $this->addFlash('danger', 'Nie można usuwać rozpoczętych zastępstw.');
+                return $this->redirectToRoute('zastepstwo');
+            }
+
+            if (!in_array("PARP_ADMIN", $this->getUser()->getRoles())
+                && !in_array("PARP_ADMIN_ZASTEPSTW", $this->getUser()->getRoles())
+                && $entity->getKogoZastepuje() != $this->getUser()->getUsername()
+            ) {
                 $this->addFlash('warning', 'Nie masz uprawnień do usunięcia nie swoich zastępstw.');
             } else {
                 $entityManager->remove($entity);
@@ -285,7 +368,7 @@ class ZastepstwoController extends Controller
         return $this->createFormBuilder()
             ->setAction($this->generateUrl('zastepstwo_delete', array('id' => $id)))
             ->setMethod('DELETE')
-            ->add('submit', SubmitType::class, array('label' => 'Skasuj Zastępstwo','attr' => array('class' => 'btn btn-danger' )))
+            ->add('submit', SubmitType::class, array('label' => 'Skasuj zastępstwo','attr' => array('class' => 'btn btn-danger' )))
             ->getForm()
         ;
     }
